+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Fix type for `headerLink` permalink options. ([#108])
 
 ## [8.6.0] - 2021-04-08
 * Add a `getTokensText` option to customize how we extract the title